Put TypingIndicator and LastMessageWidget in separate container widget
