+++ app/lib/features/chat_ui_showcase/widgets/chat_item_widget.dart
@@ -1,15 +1,12 @@
 import 'package:acter/common/providers/chat_providers.dart';
-import 'package:acter/features/chat/providers/chat_providers.dart';
 import 'package:acter/features/chat/widgets/room_avatar.dart';
+import 'package:acter/features/chat_ui_showcase/widgets/chat_item/chat_typing_lastmessage_container_widget.dart';
 import 'package:acter/features/chat_ui_showcase/widgets/chat_item/display_name_widget.dart';
 import 'package:acter/features/chat_ui_showcase/widgets/chat_item/last_message_time_widget.dart';
-import 'package:acter/features/chat_ui_showcase/widgets/chat_item/last_message_widget.dart';
 import 'package:acter/features/chat_ui_showcase/widgets/chat_item/bookmark_icon_widget.dart';
 import 'package:acter/features/chat_ui_showcase/widgets/chat_item/mute_icon_widget.dart';
-import 'package:acter/features/chat_ui_showcase/widgets/chat_item/typing_indicator.dart';
 import 'package:acter/features/chat_ui_showcase/widgets/chat_item/unread_count_widget.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_chat_types/flutter_chat_types.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class ChatItemWidget extends ConsumerWidget {
@@ -55,7 +52,6 @@ class ChatItemWidget extends ConsumerWidget {
   Widget _buildChatTitle(BuildContext context) {
     return Row(
       children: [
-        //Show chat title
         Expanded(child: DisplayNameWidget(roomId: roomId)),
         SizedBox(width: 12),
         LastMessageTimeWidget(roomId: roomId),
@@ -64,24 +60,13 @@ class ChatItemWidget extends ConsumerWidget {
   }
 
   Widget _buildChatSubtitle(BuildContext context, WidgetRef ref) {
-    final typingUsers = _getTypingUsers(ref);
     return Row(
       children: [
-        Expanded(
-          child:
-              typingUsers.isNotEmpty
-                  ? TypingIndicator(roomId: roomId)
-                  : LastMessageWidget(roomId: roomId),
-        ),
+        Expanded(child: ChatTypingLastMessageContainerWidget(roomId: roomId)),
         BookmarkIconWidget(roomId: roomId),
         MuteIconWidget(roomId: roomId),
         UnreadCountWidget(roomId: roomId),
       ],
     );
-  }
-
-  List<User> _getTypingUsers(WidgetRef ref) {
-    final users = ref.watch(chatTypingEventProvider(roomId)).valueOrNull;
-    return users ?? [];
   }
 }
